Fetching all api data correctly
