+++ frontend/package.json
@@ -8,7 +8,9 @@
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^1.3.4",
+    "geo-point": "^2.0.2",
     "google": "^2.1.0",
+    "latlon-geohash": "^2.0.0",
     "react": "^18.2.0",
     "react-datepicker": "^4.10.0",
     "react-dom": "^18.2.0",
