--- frontend/src/api/index.js
@@ -1,34 +1,0 @@
-import axios from "axios";
-import React, { useState } from "react";
-
-const URL = 'https://travel-advisor.p.rapidapi.com/restaurants/list-by-latlng';
-
-
-function getOptions(coords) {
-    return ({ method: 'GET',
-  params: {
-    latitude: coords.lat,
-    longitude: coords.lng,
-    limit: '30',
-    currency: 'USD',
-    distance: '2',
-    open_now: 'false',
-    lunit: 'km',
-    lang: 'en_US'
-  },
-  headers: {
-    'X-RapidAPI-Key': '5ae6788d7emshcd55b6020afdd20p1e2922jsn19c127ff040e',
-    'X-RapidAPI-Host': 'travel-advisor.p.rapidapi.com'
-  }});
-}
-
-export const getPlaceData = async (coords) => {
-    var options = getOptions(coords)
-    console.log(coords)
-    try {
-        const { data: { data } } = await axios.get(URL, options);
-        return data;
-    }catch (error){
-        console.log("Failed to get api data", error)
-    }
-}
